Fix Time difference logic when time is from next day
The issue was that whenever time was getting updates as per 10 second delay to show latest time in mins for the departure, the conversion was assumed to happen from hh:mm a , but it does not have any timezone or day info. therefore we need originUtcTime to use that value for calculating difference between time using date info as well.